requirements.txt
PyQt6==6.7.0
pynput==1.7.6
pyautogui==0.9.53
opencv-python==4.8.1.78
Pillow==12.1.1
numpy<2
pywin32